+++ test/test-helpers.js
@@ -8,7 +8,7 @@ import {Player, mergeOptions} from 'video.js';
 
 import adapter from 'webrtc-adapter';
 
-import {LIBVORBISJS, RECORDERJS, LAMEJS, OPUSRECORDER} from '../src/js/engine/record-engine.js';
+import {LIBVORBISJS, RECORDERJS, LAMEJS, OPUSRECORDER, VMSG} from '../src/js/engine/record-engine.js';
 import {TSEBML} from '../src/js/engine/convert-engine.js';
 
 const TestHelpers = {
@@ -131,6 +131,11 @@ const TestHelpers = {
 
             case RECORDERJS:
                 recordPluginOptions.audioEngine = RECORDERJS;
+                break;
+
+            case VMSG:
+                recordPluginOptions.audioEngine = VMSG;
+                recordPluginOptions.audioWorkerURL = '/base/node_modules/vmsg/vmsg.wasm';
                 break;
 
             default: